add ascendcl readme and update ascendc readme
