+++ python/nndeploy/test/op/test_add_ascendcl.py
@@ -11,8 +11,8 @@ class TestAddOp(unittest.TestCase):
     def test_add(self):
         input_shape = [32, 32, 64, 64]
 
-        np_input1 = np.random.random(input_shape).astype(np.float32)
-        np_input2 = np.random.random(input_shape).astype(np.float32)
+        np_input1 = np.random.random(input_shape).astype(np.float16)
+        np_input2 = np.random.random(input_shape).astype(np.float16)
 
         torch_result = torch.add(
             torch.tensor(np_input1),
@@ -22,7 +22,12 @@ class TestAddOp(unittest.TestCase):
         input1 = create_tensor_from_numpy(np_input1)
         input2 = create_tensor_from_numpy(np_input2)
 
-        nndeploy_result = F.add(input1, input2)
+        ascend_input1 = input1.to(nndeploy.base.DeviceType("ascendcl")) 
+        ascend_input2 = input2.to(nndeploy.base.DeviceType("ascendcl"))
+
+        ascend_result = F.add(ascend_input1, ascend_input2)
+
+        nndeploy_result = ascend_result.to(nndeploy.base.DeviceType("cpu"))
 
         self.assertTrue(
             np.allclose(
